+++ CHANGELOG.md
@@ -8,18 +8,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+
 * Added pluggable function `trimesh_slice` in `compas_rhino`.
+* Added equality comparison for pointclouds.
+* Added `compas.data.is_sequence_of_uint`.
 
 ### Changed
+
 * `compas.robots.Axis` is now normalized upon initialization.
 * Fixed a bug in `compas.numerical.dr_numpy` when using numpy array as inputs.
 * Allowed for varying repository file structures in `compas.robots.GithubPackageMeshLoader`.
+* Fixed data schema of `compas.geometry.Polyline`, `compas.geometry.Polygon`, `compas.geometry.Pointcloud`.
 
 ### Fixed
+
 * Fixed `Configuration.from_data` to be backward-compatible with JSON data generated before `compas 1.3.0`.
 
 ### Removed
-
 
 ## [1.7.1] 2021-06-14